api/requirements.txt
SQLAlchemy==2.0.25
psycopg2==2.9.9
Flask==2.3.2
Flask-Cors==4.0.0
gunicorn==21.1.0
PyYAML==6.0.1
jsonschema==4.17.3
httpx[http2,brotli]==0.24.1
python-jose[cryptography]==3.3.0
python_dotenv==1.0.0
praw==7.7.1
pmaw==3.0.0
Mastodon.py==1.8.1
Markdown==3.5.2
mdx-linkify==2.1
atproto==0.0.40
numpy==1.26.4
auth0-python==4.6.1